Fix merged value retrieval
Previously when retrieving a TOML table value which had both children
that were modified through IPC and children from the configuration file,
the returned value would only include the runtime overrides.

To ensure we don't just retrieve either the runtime or configuration
file value when an item is requested, we now keep a merged tree around
which represents the combination of configuration file and runtime
values.

To avoid excessive processing, this tree is only fully merged whenever
the configuration file is reloaded. For runtime overrides only the
affected value is modified directly in both the runtime and merged tree.

Additionally a new `reset` method was introduced to replace the existing
`set(…, None)` method for both `Config` and `Ipc`, which avoids having
to pass `Some` whenever trying to set a value. This also fixes a type
error during deserialization caused by `set(…, None)` always setting
removed values to an empty table, rather than removing the field
entirely.