Switch to trait-based event dispatch
The previous dispatch system was using a fixed channel to dispatch
events, which had the disadvantage of having to rely on STD channels,
making integration into some event loops difficult.

Instead of always dispatching to a channel, we now allow supplying a
struct which implements a trait to dispatch events. This can also be
implemented on top of a channel, but allows for more flexibility.