Start monitor on initial config write
Previously if configory was started up with no configuration file
present, it would never start the monitor and thus fail to deliver
configuration updates after calling `Config::persist`.

To allow persistence, the `persist` function has been moved to the
`Manager`, which allows dynamically starting the monitor when `persist`
is called for the first time with no existing watcher present.

This also fixes an issue where the monitoring thread would be started
with an empty list of files to watch, if the file didn't exist on
startup.